MA-8852 [Android_Imx8dv]Bi-weekly build: build failed in av.git on sabreauto-8v.
Failed log:
frameworks/av/media/libstagefright/FslExtractor.cpp:2493:79: error: cannot initialize a parameter of type 'size_t *' (aka 'unsigned long *') with an rvalue of type 'uin
        CHECK(trackMeta->findData(kKeyCryptoKey, &type, (const void**)&keyId, &keySize));
                                                                                      ^~~~~~~~

For sabreauto-8dv,unsigned long(64 bit) is not match with unsigned int(32 bit).

Reviewed-on: http://androidsource.ap.freescale.net/project/1196

+++ media/libstagefright/FslExtractor.cpp
@@ -2485,7 +2485,7 @@ status_t FslExtractor::SetMkvCrpytBufferInfo(TrackInfo *pInfo, MediaBuffer *buf)
 
         uint32 type = 0;
         uint8 *keyId = NULL;
-        uint32 keySize = 0;
+        size_t keySize = 0;
 
         memcpy(ctrCounter, buffer_ptr + 1, 8);
